% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{compute.change.variable}
\alias{compute.change.variable}
\title{Compute change variables}
\usage{
compute.change.variable(climResponse, lOpt, lDim, iCpredContUnique, iCpredCont)
}
\arguments{
\item{climResponse}{output from \code{\link{extract.climate.response}}}

\item{lOpt}{list of options, returned by \code{\link{QUALYPSOSS.check.option}}}

\item{lDim}{list of dimensions}

\item{iCpredContUnique}{index in \code{1:nP} indicating the reference continuous predictor 
for the computation of change variables.}

\item{iCpredCont}{index in \code{1:nT} indicating the reference period (reference period) 
for the computation of change variables.}
}
\value{
list with the following fields:
\itemize{
  \item \strong{phiStar.MCMC}: MCMC draws of climate change response
  \item \strong{etaStar.MCMC}: MCMC draws of deviation from the climate change response
}
}
\description{
Compute change variables
}
\author{
Guillaume Evin
}
