% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{formatQUALYPSSoutput}
\alias{formatQUALYPSSoutput}
\title{formatQUALYPSSoutput}
\usage{
formatQUALYPSSoutput(
  lOpt,
  lDim,
  lScen,
  ANOVA.step1,
  ANOVA.step2,
  ANOVA.step3,
  climResponse,
  change.variable
)
}
\arguments{
\item{lOpt}{list of options, returned by \code{\link{QUALYPSOSS.check.option}}}

\item{lDim}{list of dimensions}

\item{lScen}{list of scenario characteristics, output from \code{\link{QUALYPSOSS.process.scenario}}}

\item{ANOVA.step1}{list provided by \code{\link{QUALYPSOSS.ANOVA.step1}}}

\item{ANOVA.step2}{list provided by \code{\link{QUALYPSOSS.ANOVA.step2}}}

\item{ANOVA.step3}{list provided by \code{\link{QUALYPSOSS.ANOVA.step3}}}

\item{climResponse}{list containing phi, eta, provided by \code{\link{extract.climate.response}}}

\item{change.variable}{list containing phiStar, etaStar, provided by \code{\link{compute.change.variable}}}
}
\value{
list with the following fields:

\itemize{
  \item \strong{POINT}: list containing the mean estimate of different quantities: \code{RESIDUALVAR} 
  (residual variability), \code{INTERNALVAR} (internal variability), \code{GRANDMEAN} (grand mean for all time
  steps), \code{MAINEFFECT} (list with one item per discrete predictor \code{i}, containing matrices \code{nT} x 
  \code{nEffi}, where \code{nEffi} is the number of possible values for the discrete predictor \code{i}).
  \code{EFFECTVAR}, uncertainty related to the different main effect, \code{TOTVAR} Total variability,
  \code{DECOMPVAR}, decomposition of the total variability (percentages) for the different components,
  \code{CONTRIB_EACH_EFFECT}, contribution of each individual effects (percentages) to the corr. effect uncertainty.
  \item \strong{BAYES}: list containing quantiles of different estimated quantities, listed in \strong{POINT}.
  \item \strong{MCMC}: MCMC draws for the different quantities.
}
}
\description{
formatQUALYPSSoutput
}
\author{
Guillaume Evin
}
