% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.det.AR1}
\alias{get.det.AR1}
\title{get.det.AR1 
return the determinant of the matrix provided by  \code{\link{get.matrix.AR1}}}
\usage{
get.det.AR1(nP, rho, nMO)
}
\arguments{
\item{nP}{number of continuous predictors (e.g. future times)}

\item{rho}{AR(1) correlation parameter in (-1,1)}

\item{nMO}{number of possible simulation chains (missing and non-missing)}
}
\value{
determinant of the AR1 matrix
}
\description{
get.det.AR1 
return the determinant of the matrix provided by  \code{\link{get.matrix.AR1}}
}
\author{
Guillaume Evin
}
