% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.det.KMS}
\alias{get.det.KMS}
\title{get.det.KMS 
return the determinant of the KMS matrix}
\usage{
get.det.KMS(nP, rho)
}
\arguments{
\item{nP}{number of continuous predictors (e.g. future times)}

\item{rho}{AR(1) correlation parameter in (-1,1)}
}
\value{
determinant of the KMS matrix
}
\description{
get.det.KMS 
return the determinant of the KMS matrix
}
\author{
Guillaume Evin
}
