% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.logdet.W}
\alias{get.logdet.W}
\title{get.logdet.W 
Return the logarithm of the determinant of the matrix W}
\usage{
get.logdet.W(weight.hetero, nMO, nP, rho)
}
\arguments{
\item{weight.hetero}{output of \code{\link{get.vec.weight.hetero}}}

\item{nMO}{number of possible simulation chains (missing and non-missing)}

\item{nP}{number of continuous predictors (e.g. future times)}

\item{rho}{AR(1) correlation parameter in (-1,1)}
}
\value{
logarithm of the determinant of the matrix W
}
\description{
get.logdet.W 
Return the logarithm of the determinant of the matrix W
}
\author{
Guillaume Evin
}
