% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.matrix.AR1}
\alias{get.matrix.AR1}
\title{get.matrix.AR1 
return the matrix of AR(1) correlations corresponding to the entire ensemble}
\usage{
get.matrix.AR1(nP, rho, nMO)
}
\arguments{
\item{nP}{number of continuous predictors (e.g. future times)}

\item{rho}{AR(1) correlation parameter in (-1,1)}

\item{nMO}{number of possible simulation chains (missing and non-missing)}
}
\value{
C matrix n x n of AR(1) correlations where code{n} is the total number of predictions 
(all the predictions for all the possible simulation chains)
}
\description{
get.matrix.AR1 
return the matrix of AR(1) correlations corresponding to the entire ensemble
}
\author{
Guillaume Evin
}
