% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.matrix.hetero.inv}
\alias{get.matrix.hetero.inv}
\title{get.matrix.hetero.inv 
returns the inverse of the matrix of weights for the computation of heteroscedastic errors
corresponding to the entire ensemble}
\usage{
get.matrix.hetero.inv(weight.hetero, nMO)
}
\arguments{
\item{weight.hetero}{output of \code{\link{get.vec.weight.hetero}}}

\item{nMO}{number of possible simulation chains (missing and non-missing)}
}
\value{
inverse matrix n x n of weights where code{n} is the total number of predictions 
(all the predictions for all the possible simulation chains)
}
\description{
get.matrix.hetero.inv 
returns the inverse of the matrix of weights for the computation of heteroscedastic errors
corresponding to the entire ensemble
}
\author{
Guillaume Evin
}
