% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.matrix.KMS}
\alias{get.matrix.KMS}
\title{get.matrix.KMS
Return the square Kac-Murdoch-Szego matrix for a rho correlation and n lines/colums}
\usage{
get.matrix.KMS(n, rho)
}
\arguments{
\item{n}{nummber of lines/columns of the square matrix}

\item{rho}{correlation parameter in [0,1]}
}
\value{
n x n Kac-Murdock-Szego matrix
}
\description{
get.matrix.KMS
Return the square Kac-Murdoch-Szego matrix for a rho correlation and n lines/colums
}
\references{
Kac, M., W. L. Murdock, and G. Szego. 1953. 'On the Eigen-Values of Certain Hermitian Forms'
Journal of Rational Mechanics and Analysis 2: 767-800.
}
\author{
Guillaume Evin
}
