% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.spectral.decomp}
\alias{get.spectral.decomp}
\title{get.spectral.decomp}
\usage{
get.spectral.decomp(SIGMA)
}
\arguments{
\item{SIGMA}{reproducing kernel}
}
\value{
list with the following fields:
\itemize{
  \item \strong{Q}: Matrix of eigen vectors n x r,
  \item \strong{D}: Vector of nonzero eigen values (size r),
  \item \strong{r}: Number of nonzero eigen values (scalar).
}
}
\description{
compute different objects used for the application of Smoothing-Splines ANOVA (SS-ANOVA)
}
\author{
Guillaume Evin
}
