% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.target.logdensity.rho}
\alias{get.target.logdensity.rho}
\title{get.target.density.rho 
Return the log-density of the full conditional distribution for the parameter rho}
\usage{
get.target.logdensity.rho(nFull, deltaRV, distSS, weight.hetero, nMO, nP, rho)
}
\arguments{
\item{nFull}{nP x nMO}

\item{deltaRV}{variance of the residual terms for the max value of the continuous predictor}

\item{distSS}{sum of square distances between the climate change responses and the ANOVA model}

\item{weight.hetero}{output of \code{\link{get.vec.weight.hetero}}}

\item{nMO}{number of possible simulation chains (missing and non-missing)}

\item{nP}{number of continuous predictors (e.g. future times)}

\item{rho}{AR(1) correlation parameter in (-1,1)}
}
\value{
log-density of the full conditional distribution
}
\description{
get.target.density.rho 
Return the log-density of the full conditional distribution for the parameter rho
}
\author{
Guillaume Evin
}
