% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.vec.weight.hetero}
\alias{get.vec.weight.hetero}
\title{get.vec.weight.hetero 
returns the vector of weights for the computation of heteroscedastic errors
corresponding to one simulation chain}
\usage{
get.vec.weight.hetero(nP, type.weight.hetero)
}
\arguments{
\item{nP}{length of the continuous predictor for which we want to obtain the prediction (e.g. time)
we suppose that continuous predictor is regularly spaced (e.g. 1990,2000,2010,...)}

\item{type.weight.hetero}{"constant" (homoscedastic) or "linear" (heteroscedastic)}
}
\value{
vector of square roots of weights of the same length than \code{predContUnique}
}
\description{
get.vec.weight.hetero 
returns the vector of weights for the computation of heteroscedastic errors
corresponding to one simulation chain
}
\author{
Guillaume Evin
}
