% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{get.yMCMC}
\alias{get.yMCMC}
\title{get.yMCMC}
\usage{
get.yMCMC(lOpt, lDim, lScen, change.variable)
}
\arguments{
\item{lOpt}{list of options, returned by \code{\link{QUALYPSOSS.check.option}}}

\item{lDim}{list of dimensions}

\item{lScen}{list of scenario characteristics, output from \code{\link{QUALYPSOSS.process.scenario}}}

\item{change.variable}{output from \code{\link{compute.change.variable}} containing MCMC draws 
of climate change response}
}
\value{
strong{yMCMC}: matrix \code{nMCMC} x \code{nFull} of climate responses
}
\description{
Get matrix \code{nMCMC} x \code{nFull} of climate responses where \code{nMCMC}
is the number of MCMC draws and \code{nFull} is the number of possible combinations of 
predictors (discrete AND continuous),
}
\author{
Guillaume Evin
}
