% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{plotQUALYPSOSSClimateChangeResponse}
\alias{plotQUALYPSOSSClimateChangeResponse}
\title{plotQUALYPSOSSClimateChangeResponse}
\usage{
plotQUALYPSOSSClimateChangeResponse(QUALYPSOSSOUT, lim = NULL,
  col = NULL, xlab = "Years", ylab = expression(phi^{     star }),
  ...)
}
\arguments{
\item{QUALYPSOSSOUT}{output from \code{\link{QUALYPSOSS}}}

\item{lim}{y-axis limits (default is NULL)}

\item{col}{color for the lines}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot climate change responses.
}
\author{
Guillaume Evin
}
