% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{plotQUALYPSOSSClimateResponse}
\alias{plotQUALYPSOSSClimateResponse}
\title{plotQUALYPSOSSClimateResponse}
\usage{
plotQUALYPSOSSClimateResponse(
  QUALYPSOSSOUT,
  lim = NULL,
  col = NULL,
  xlab = "Years",
  ylab = expression(phi),
  ...
)
}
\arguments{
\item{QUALYPSOSSOUT}{output from \code{\link{QUALYPSOSS}}}

\item{lim}{y-axis limits (default is NULL)}

\item{col}{color for the lines}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot climate responses.
}
\author{
Guillaume Evin
}
