% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{plotQUALYPSOSSeffect}
\alias{plotQUALYPSOSSeffect}
\title{plotQUALYPSOSSeffect}
\usage{
plotQUALYPSOSSeffect(
  QUALYPSOSSOUT,
  iEff,
  CIlevel = c(0.025, 0.975),
  lim = NULL,
  col = 1:20,
  xlab = "Continuous predictor",
  ylab = "Effect",
  addLegend = TRUE,
  ...
)
}
\arguments{
\item{QUALYPSOSSOUT}{output from \code{\link{QUALYPSOSS}}}

\item{iEff}{index of the main effect to be plotted in \code{QUALYPSOSSOUT$listScenario$predDiscreteUnique}}

\item{CIlevel}{probabilities for the credible intervals, default is equal to \code{c(0.025,0.975)}}

\item{lim}{y-axis limits (default is NULL)}

\item{col}{colors for each effect}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{addLegend}{if TRUE, a legend is added}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot prediction of ANOVA effects for one main effect. By default, we plot we plot the credible intervals corresponding to a probability 0.95.
}
\author{
Guillaume Evin
}
