% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{plotQUALYPSOSSgrandmean}
\alias{plotQUALYPSOSSgrandmean}
\title{plotQUALYPSOSSgrandmean}
\usage{
plotQUALYPSOSSgrandmean(QUALYPSOSSOUT, CIlevel = c(0.025, 0.975),
  lim = NULL, col = "black", xlab = "Continuous predictor",
  ylab = "Grand mean", addLegend = T, ...)
}
\arguments{
\item{QUALYPSOSSOUT}{output from \code{\link{QUALYPSOSS}}}

\item{CIlevel}{probabilities for the credible intervals, default is equal to \code{c(0.025,0.975)}}

\item{lim}{y-axis limits (default is NULL)}

\item{col}{color for the overall mean and the credible interval}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{addLegend}{if TRUE, a legend is added}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot prediction of grand mean ensemble. By default, we plot the credible interval corresponding to a probability 0.95.
}
\author{
Guillaume Evin
}
