% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{plotQUALYPSOSSTotalVarianceDecomposition}
\alias{plotQUALYPSOSSTotalVarianceDecomposition}
\title{plotQUALYPSOSSTotalVarianceDecomposition}
\usage{
plotQUALYPSOSSTotalVarianceDecomposition(QUALYPSOSSOUT, col = c("orange",
  "yellow", "cadetblue1", "blue1", "darkgreen", "darkgoldenrod4",
  "darkorchid1"), xlab = "Continuous predictor",
  ylab = "\% Total Variance", addLegend = TRUE, ...)
}
\arguments{
\item{QUALYPSOSSOUT}{output from \code{\link{QUALYPSOSS}}}

\item{col}{colors for each source of uncertainty, the first two colors corresponding to internal variability and residual variability, respectively}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{addLegend}{if TRUE, a legend is added}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot fraction of total variance explained by each source of uncertainty.
}
\author{
Guillaume Evin
}
