% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{predGlobTemp}
\alias{predGlobTemp}
\title{Annual average of global temperatures simulated by different CMIP5 GCMs at
the planetary scale for the period 1971-2099}
\format{
matrix 129 years x 20 scenarios
}
\usage{
data(predGlobTemp)
}
\description{
Annual average of global temperatures simulated by different CMIP5 GCMs at
the planetary scale for the period 1971-2099
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
