% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{predTime}
\alias{predTime}
\title{Years 1971-2099 repeated for the 20 scenarios}
\format{
matrix 129 years x 20 scenarios
}
\usage{
data(predTime)
}
\description{
Years 1971-2099 repeated for the 20 scenarios
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
