% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{predTimeUnique}
\alias{predTimeUnique}
\title{Equally spaced vector of years over the period 1971-2099}
\format{
vector of length 13
}
\usage{
data(predTimeUnique)
}
\description{
Equally spaced vector of years over the period 1971-2099
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
