% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{QUALYPSOSS.ANOVA}
\alias{QUALYPSOSS.ANOVA}
\title{QUALYPSOSS.ANOVA}
\usage{
QUALYPSOSS.ANOVA(lOpt, yMCMC, RK)
}
\arguments{
\item{lOpt}{list of options, returned by \code{\link{QUALYPSOSS.check.option}}}

\item{yMCMC}{array \code{nMCMC} x \code{nFull} of climate change responses}

\item{RK}{large object containing the reproducing kernels, returned by \code{\link{QUALYPSOSS.get.RK}}}
}
\value{
list containing diverse information aboutwith the following fields:
\itemize{
  \item \strong{g}: Smooth effects \code{g}: array \code{n} x \code{nFull} x \code{L} where
  \code{nFull} is the number of possible combinations of predictors (discrete AND continuous),
  \item \strong{lambda}: Smoothing parameters: matrix \code{n} x \code{L},  
  \item \strong{sigma2}: Residual variance: vector of length \code{n},
  \item \strong{MCMC.list}: list containing previous objects, for records (according to the option \code{returnMCMC}).
}
}
\description{
SSANOVA decomposition of the ensemble of climate change responses using a Bayesian approach.
The different fields of the returned list contain \code{n} samples from the posterior distributions
of the different inferred quantities.
}
\author{
Guillaume Evin
}
