% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{QUALYPSOSS.check.option}
\alias{QUALYPSOSS.check.option}
\title{QUALYPSOSS.check.option}
\usage{
QUALYPSOSS.check.option(listOption)
}
\arguments{
\item{listOption}{list of options}
}
\value{
List containing the complete set of options.
}
\description{
Check if input options provided in \code{\link{QUALYPSOSS}} are valid and assigned default values if missing.
}
\author{
Guillaume Evin
}
