% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{QUALYPSOSS.get.RK}
\alias{QUALYPSOSS.get.RK}
\title{QUALYPSOSS.get.RK}
\usage{
QUALYPSOSS.get.RK(X, nK)
}
\arguments{
\item{X}{matrix of predictors}

\item{nK}{number of discrete predictors}
}
\value{
strong{RK}: list containing the reproducing kernels, obtained using spectral decomposition
}
\description{
Get reproducing kernel for each discrete predictor
}
\author{
Guillaume Evin
}
