% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{QUALYPSOSS.process.scenario}
\alias{QUALYPSOSS.process.scenario}
\title{QUALYPSOSS.process.scenario}
\usage{
QUALYPSOSS.process.scenario(scenAvail, predContUnique)
}
\arguments{
\item{scenAvail}{matrix of scenario characteristics \code{nS} x \code{nK}.}

\item{predContUnique}{(optional) unique values of continuous predictors.}
}
\value{
list containing diverse information aboutwith the following fields:
\itemize{
  \item \strong{scenAvail}: Record first argument of the function,
  \item \strong{predContUnique}: Record second argument of the function,  
  \item \strong{XFull}: data.frame with all possible combinations of predictors (continuous AND discrete),
  \item \strong{nFull}: number of rows of \code{XFull},
  \item \strong{nK}: Number of columns of \code{ScenAvail} (i.e. number of discrete predictors),   
  \item \strong{predDiscreteUnique}: List containing possible values for each discrete predictor.
}
}
\description{
compute different objects used for the application of Smoothing-Splines ANOVA (SS-ANOVA),
these objects being processed outputs of the scenario characteristics
}
\author{
Guillaume Evin
}
