% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{reproducing.kernel}
\alias{reproducing.kernel}
\title{reproducing.kernel}
\usage{
reproducing.kernel(x, y = NULL, type, typeRK = "Cheng")
}
\arguments{
\item{x}{vector of predictors (continuous or discrete)}

\item{y}{vector of predictors (continuous or discrete)}

\item{type}{'continuous' or 'discrete'}

\item{typeRK}{type of reproducing kernels: c('Cheng','Gu','Gaussian')}
}
\value{
matrix n x n
}
\description{
see par 2.3 in Cheng and Speckman
}
\author{
Guillaume Evin
}
