% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{vecYears}
\alias{vecYears}
\title{vecYears gives the years corr. to Y, i.e. from 1971 to 2099}
\format{
vectors of length 129
}
\usage{
data(vecYears)
}
\description{
vecYears gives the years corr. to Y, i.e. from 1971 to 2099
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
