% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSOSS.r
\name{QUALYPSOSS.get.RK}
\alias{QUALYPSOSS.get.RK}
\title{QUALYPSOSS.get.RK}
\usage{
QUALYPSOSS.get.RK(X, nK, nCluster)
}
\arguments{
\item{X}{matrix of predictors}

\item{nK}{number of discrete predictors}

\item{nCluster}{number of clusters used to compute the reproducing kernels}
}
\value{
strong{RK}: list containing the reproducing kernels, obtained using spectral decomposition
}
\description{
Get reproducing kernel for each discrete predictor
}
\author{
Guillaume Evin
}
